web-animations/animation-node/animation-node-next-sibling.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation node nextSibling attribute test</title>
<meta name="assert" content="The next sibling of this animation node">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationNode-nextSibling">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var node = newAnimation(createDiv(this));

    assert_equals(node.nextSibling, null, 'AnimationNode.nextSibling should be null');
}, 'AnimationNode.nextSibling is null if animation node does not have next sibling');

test(function() {
    var node = new AnimationGroup([]);

    assert_equals(node.nextSibling, null, 'AnimationNode.nextSibling should be null');
}, 'AnimationNode.nextSibling is null if animation node does not have next sibling. ' +
    'Test AnimationGroup');

test(function() {
    var node = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node]);

    assert_equals(node.nextSibling, null, 'AnimationNode.nextSibling ' +
        'should be null');
}, 'AnimationNode.nextSibling should be null if animation group contains only one node');

test(function() {
    var node = new AnimationGroup([]);
    var animationGroup = new AnimationGroup([node]);

    assert_equals(node.nextSibling, null, 'AnimationNode.nextSibling ' +
        'should be null');
}, 'AnimationNode.nextSibling should be null if animation group contains only one node. ' +
    'Test AnimationGroup');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node1, node2]);

    assert_equals(node1.nextSibling, node2, 'AnimationNode.previousSibling should return ' +
        'next sibling of this animation node');
    assert_equals(node2.nextSibling, null, 'AnimationNode.previousSibling should be null');
}, 'AnimationNode.nextSibling returns next sibling of this animation node');

test(function() {
    var node1 = new AnimationGroup([]);
    var node2 = new AnimationGroup([]);
    var animationGroup = new AnimationGroup([node1, node2]);

    assert_equals(node2.nextSibling, null, 'AnimationNode.nextSibling should be null');
    assert_equals(node1.nextSibling, node2, 'AnimationNode.nextSibling should return ' +
        'next sibling of this animation node');
}, 'AnimationNode.nextSibling returns next sibling of this animation node. Test ' +
    'AnimationGroup');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = newAnimation(createDiv(this));
    var node4 = newAnimation(createDiv(this));
    var node5 = newAnimation(createDiv(this));
    var node6 = newAnimation(createDiv(this));
    var node7 = new AnimationGroup([node3, node4]);
    var node8 = new AnimationGroup([node5, node6]);
    var node9 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node1, node2, node7, node8, node9]);

    assert_equals(node1.nextSibling, node2, 'AnimationNode.nextSibling should return ' +
        'next sibling for the first node in the group');
    assert_equals(node2.nextSibling, node7, 'AnimationNode.nextSibling should return ' +
        'next sibling of this animation node');
    assert_equals(node3.nextSibling, node4, 'AnimationNode.nextSibling should return ' +
        'next sibling of this animation node (first node in the nested group)');
    assert_equals(node4.nextSibling, null, 'AnimationNode.nextSibling should be null ' +
        'for the last node in the nested group');
	assert_equals(node5.nextSibling, node6, 'AnimationNode.nextSibling should return ' +
        'next sibling of this animation node (first node in the second nested group)');
    assert_equals(node6.nextSibling, null, 'AnimationNode.nextSibling should be null ' +
        'for the last node in the second nested group');
    assert_equals(node7.nextSibling, node8, 'AnimationNode.nextSibling should return ' +
        'next sibling of this animation node (first nested group)');
    assert_equals(node8.nextSibling, node9, 'AnimationNode.nextSibling should return ' +
        'next sibling of this animation node (second nested group)');
    assert_equals(node9.nextSibling, null, 'AnimationNode.nextSibling should return ' +
        'null (the last node)');
}, 'AnimationNode.previousSibling returns previous sibling of this animation node. Test ' +
    'tree structure');
</script>
</body>
